Moves getItem functions to e2eHelpers for reusability
